try fix compilation warning on mac
copy2shm.c:244:29: warning: format specifies type 'long' but the argument has type 'off_t' (aka 'long long') [-Wformat]

+++ src/copy2shm.c
@@ -240,8 +240,8 @@ SEXP allocate_from_shm(SEXP name, SEXP type, SEXP length, SEXP size,
 
   if (sb.st_size != pages * pagesize) {
     close(fd);
-    error("file backing shm object is of wrong size; expected: %ld bytes, actual: %ld bytes",
-          pages * pagesize, sb.st_size);
+    error("file backing shm object is of wrong size; expected: %ld bytes, actual: %jd bytes",
+          pages * pagesize, (intmax_t) sb.st_size);
   }
 #endif
